Improve end-of-stream handling.
Reading from the stream may fail when the end of stream is
reached. Since this is a valid reason for the read to fail, the code
must properly handle it. Introduce the necessary control flow to
recognize valid end-of-stream scenarios and handle them by returning
to caller w/o raising an error.