Generalize handling of decoded frames.
To decide what happens with a decoded frame, the `parse_mpeg` function
now accepts a new trait parameter. `parse_mpeg` calls the trait once
for each decoded frame. The actual logic of what happens with the
fame, e.g., displaying it or writing the frame to disk as an image, is
encapsulated in the object implementing the trait.